Put more stuff into extension methods
